Lab9: Code fixes
Dear students,

We were made aware of three issues concerning Lab 9 and have made a fix that we
have now pushed to the git repository. Here is a summary of the issues:

1) State reset of legs: When a new set of parameters would be run, the joints of
the body would be reset correctly but not those of the legs. This would cause
the legs to rewind (multiple rotations) back to the zeros position. This has
been fixed in reset.py.

2) Velocity logging: When working with position controllers in Webots,
motor.getVelocity returns the maximum velocity of the motor instead of the
velocity. We have fixed this by differentiating the position of the motor at
each timestep. The changes were applied in experiment_logger.py.

3) The legs were not logged: We have now added the legs joints logs when running
Webots simulations in cmc_robot.py.

You can fix these issues by simply running "git pull" in your local copy of the
repository if you have cloned it with git. Otherwise, you also have the
possibility to directly replace the old files reset.py, experiment_logger.py and
cmc_robot.py provided you have not made any changes to them. If you decide to
copy the files and have made changes to them, please watch out for those changes
and apply them to the new file.

We are sorry for the inconvenience and remain available in case of any problem
when updating.

Good luck with the end of the project,
Your CMC assistants